cloudfunctions/requirements.txt
functions-framework
google-cloud-firestore
numpy
flask
